clear for foc
